+++ packages/app_pick_crop_image_flutter/lib/src/platform_web.dart
@@ -6,31 +6,6 @@ import 'dart:typed_data';
 import 'package:image_picker/image_picker.dart';
 import 'package:tekartik_app_pick_crop_image_flutter/src/pick_image_web.dart';
 import 'package:tekartik_app_pick_crop_image_flutter/src/picked_file.dart';
-import 'package:web/web.dart' as web;
-
-/// Save image file.
-Future<void> saveImageFile({
-  required Uint8List bytes,
-  required String mimeType,
-  required String filename,
-}) async {
-  // prepare
-  final blob = web.Blob([bytes.toJS].toJS);
-  final url = web.URL.createObjectURL(blob);
-  final anchor = web.document.createElement('a') as web.HTMLAnchorElement
-    ..href = url
-    ..style.display = 'none'
-    ..download = filename;
-  web.document.body!.children.add(anchor);
-
-  // download
-  anchor.click();
-
-  // cleanup
-  anchor.remove();
-
-  web.URL.revokeObjectURL(url);
-}
 
 final _picker = ImagePicker();